apps/api/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15.0
python-dotenv>=1.0.1
prisma>=0.10.0
django-filter>=23.4
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1